prepare secondary modules
