+++ AdvisedJunit4/src/test/java/org/ehoffman/aop/objectfactory/TestSimpleObjectFactory.java
